+++ nix/ledgerblue.nix
@@ -1,4 +1,4 @@
-{ pkgs ? import (import ./dep/ledger-platform/thunk.nix + "/dep/nixpkgs") {}
+{ pkgs ? import (import ./dep/alamgu/thunk.nix + "/dep/nixpkgs") {}
 , ...
 }: